% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-acct_type_balance.R
\docType{data}
\name{acct_type_balance}
\alias{acct_type_balance}
\title{Data on various bank accounts}
\format{
A tibble with 50 rows and 2 variables:
\describe{
\item{bank_account_type}{Type of bank account: "Checking" or "Savings" }
\item{balance}{Account balance in $}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
acct_type_balance
}
\description{
Account balances in checking and savings accounts
}
\keyword{datasets}
