% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-advertising_sales_channel.R
\docType{data}
\name{advertising_sales_channel}
\alias{advertising_sales_channel}
\title{Advertising and sales dataset}
\format{
A tibble with 400 rows and 3 variables:
\describe{
\item{channel}{Sales channel: "Search Ads" or "Retail promo" }
\item{advertising_spend_k}{Weekly advertising expenditure in $'000}
\item{weekly_sales_k}{Weekly units sold units '000l}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
advertising_sales_channel
}
\description{
Sales expenditure and weekly sales by sales channel
}
\keyword{datasets}
