% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-age_income_gender_buy.R
\docType{data}
\name{age_income_gender_buy}
\alias{age_income_gender_buy}
\title{Data purchases of a product}
\format{
A tibble with 100 rows and 4 variables:
\describe{
\item{age}{Age of the person in years}
\item{income}{Annual income of person in $}
\item{gender}{Gender of the person: "F", "M"}
\item{buy_status}{Whether they bought the product or not: "Buy", "No buy"}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
age_income_gender_buy
}
\description{
Data on purchases of a product by various people
}
\keyword{datasets}
