% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-BostonHousing.R
\docType{data}
\name{BostonHousing}
\alias{BostonHousing}
\title{Data on home prices in Boston neighborhoods}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 506 rows and 14 columns.
}
\source{
\url{https://www.kaggle.com/c/boston-housing}
}
\usage{
BostonHousing
}
\description{
Data on several Boston neighborhhods showing home prices and many other details about the homes andthe neighborhoods
}
\keyword{datasets}
