% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-employees.R
\docType{data}
\name{employees}
\alias{employees}
\title{Data about employees in a company}
\format{
A tibble with 5 rows and 6 variables:
}
\source{
Synthetic data created for teaching.
}
\usage{
employees
}
\description{
Toy dataset created to illustrate missing values in a dataset
}
\keyword{datasets}
