% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exp_demand.R
\docType{data}
\name{exp_demand}
\alias{exp_demand}
\title{Data on various bank accounts
Data showing marketing expenditure and demand for a company's products}
\format{
A tibble with 600 rows and 3 variables:
\describe{
\item{Product}{Product corresponding to the row: "X", "Y", or "Z"}
\item{Expenditure}{Marketing expenditure in $}
\item{Demand}{Demand for product in units}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
exp_demand
}
\description{
Weekly marketing expenditure and demand for various products of a company. Data created to illustrate Simpson's paradox
}
\keyword{datasets}
