% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-homes.R
\docType{data}
\name{homes}
\alias{homes}
\title{Data on various homes}
\format{
A tibble with 100 rows and 5 variables:
\describe{
\item{age}{age of home in years}
\item{floor_area_sqft}{Floor area of the home in square feet}
\item{number_of_bedrooms}{The number of bedrooms the home has}
\item{no_of_bathrooms}{The number of bathrooms the home has}
\item{price}{Price of the home in $}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
homes
}
\description{
Data showing various attributes of homes and the hime price
}
\keyword{datasets}
