% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ice_cream_paradox.R
\docType{data}
\name{ice_cream_paradox}
\alias{ice_cream_paradox}
\title{Data on consumption of ice cream in various cities}
\format{
A tibble with 500 rows and 3 variables:
\describe{
\item{avg_summe_temperature}{degrees C}
\item{population}{population of the city: number of people}
\item{ice_cream_consumption}{pounds of ice cream consumed in a year}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
ice_cream_paradox
}
\description{
Consumption of ice cream in various cities. Created to show how we need to "adjust" for some variables
}
\keyword{datasets}
