% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-kiosk.R
\docType{data}
\name{kiosk}
\alias{kiosk}
\title{Data relating number of customers visiting a beach kiosk to daily temperature}
\format{
A tibble with 12 rows and 3 variables:
\describe{
\item{temperature}{High temperature on a given day}
\item{customers}{Number of customers who visited the kiosk}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
kiosk
}
\description{
Data relating number of customers visiting a beach kiosk to daily temperature
}
\keyword{datasets}
