% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-kiosk_beach_mall_temp.R
\docType{data}
\name{kiosk_beach_mall_temp}
\alias{kiosk_beach_mall_temp}
\title{Data relating of customers visiting two kiosks to the temperature}
\format{
A tibble with 12 rows and 3 variables:
\describe{
\item{day}{D01 to D25 -- Identifies the day on which the data was gathered}
\item{customers}{Number of customers}
\item{kiosk}{Location of kiosk: Mall or Beach}
\item{temperature}{High temperature on that day}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
kiosk_beach_mall_temp
}
\description{
Data relating of customers visiting two kiosks to the temperature
}
\keyword{datasets}
