% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-kiosk_mall_beach.R
\docType{data}
\name{kiosk_mall_beach}
\alias{kiosk_mall_beach}
\title{Data on number of customers visiting two kiosks}
\format{
A tibble with 12 rows and 3 variables:
\describe{
\item{customers}{Number of customers visiting a kiosk}
\item{kiosk}{Location of kiosk: Mall or Beach}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
kiosk_mall_beach
}
\description{
Data on number of customers visiting two kiosks
}
\keyword{datasets}
