% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-movies.R
\docType{data}
\name{movies}
\alias{movies}
\title{Data on movies}
\format{
A tibble with 211 rows and 4 variables:
\describe{
\item{Movie}{Name of movie}
\item{Genre}{Genre of movie: "Horror", "Drama", ...}
\item{Budget}{Budget for the movie: $}
\item{Gross_sales}{Gross box office sales: $}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
movies
}
\description{
Data on movies showing genre
}
\keyword{datasets}
