% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-price_demand.R
\docType{data}
\name{price_demand}
\alias{price_demand}
\title{Data on prices and quantities sold of a product in a company in several sales areas}
\format{
A tibble with 500 rows and 2 variables:
\describe{
\item{price}{Price charged in the sales area}
\item{sales_qty}{Number of units sold in the sales area}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
price_demand
}
\description{
Data prices and quantities sold of a product in a company
}
\keyword{datasets}
