% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-production_cost.R
\docType{data}
\name{production_cost}
\alias{production_cost}
\title{Data on production costs at a company}
\format{
A tibble with 100 rows and 2 variables:
\describe{
\item{units_produced}{Number of units produced}
\item{total_cost}{Total production cost: $}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
production_cost
}
\description{
Data on total production costs in a company
}
\keyword{datasets}
