% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-returns_dpo.R
\docType{data}
\name{returns_dpo}
\alias{returns_dpo}
\title{Data on daily retuns of stocks}
\format{
A tibble with 100 rows and 2 variables:
\describe{
\item{daily return}{Daily retun on the stock}
\item{dpo}{Number of days of payments outstanding to suppliers}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
returns_dpo
}
\description{
Data on daily returns on stocks and their days payment outstanding to suppliers
}
\keyword{datasets}
