% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sales.R
\docType{data}
\name{sales}
\alias{sales}
\title{Data relating price to demand in various regions}
\format{
A tibble with 12 rows and 3 variables:
\describe{
\item{price}{Price charged}
\item{demand}{Number of units demanded by customers}
\item{region}{Sales region}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
sales
}
\description{
Data relating price to demand in various regions
}
\keyword{datasets}
