% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sales_advertising.R
\docType{data}
\name{sales_advertising}
\alias{sales_advertising}
\title{Data sales and advertising for two products}
\format{
A tibble with 1000 rows and 3 variables:
\describe{
\item{sales}{Sales in $}
\item{advertising}{Advertising expenses in $s}
\item{product}{Corresponding product}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
sales_advertising
}
\description{
Data SHowing various weeks and their corresponding advertising and sales
}
\keyword{datasets}
