% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-variance_example.R
\docType{data}
\name{variance_example}
\alias{variance_example}
\title{Toy data to compute variance}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 1 columns.
}
\source{
Synthetic data created for teaching.
}
\usage{
variance_example
}
\description{
Toy data to compute variance
}
\keyword{datasets}
