% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sales.R
\docType{data}
\name{sales}
\alias{sales}
\title{Data osales for a company in various regions}
\format{
A tibble with 12 rows and 3 variables:
\describe{
\item{daily price}{Price chargedk}
\item{demand}{Number of units demanded by customers}
\item{region}{Sales region}
}
}
\source{
Synthetic data created for teaching.
}
\usage{
sales
}
\description{
Data on daily returns on stocks and their days payment outstanding to suppliers
}
\keyword{datasets}
